updated the code to include date
