--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-  "python.testing.pytestArgs": ["."],
-  "python.testing.unittestEnabled": false,
-  "python.testing.pytestEnabled": true
-}
