+++ pyproject.toml
@@ -24,7 +24,7 @@ dev = [
 [tool.pytest.ini_options]
 testpaths = ["tests"]
 python_files = ["test_*.py"]
-addopts = "-v --cov=src/agrimarket"
+addopts = "-v --cov=agrimarket"
 
 [tool.ruff]
 line-length = 88
